backend/requirements.txt
fastapi
uvicorn
livekit
python-dotenv
groq
twilio
